requirements.txt
pyrofork == 2.3.61
tgcrypto-pyrofork == 1.2.7
pytz == 2024.2
numpy == 2.1.2
pydantic == 2.11.1
simplejson == 3.20.1
httpx == 0.28.1
python-dotenv